All Components: removed extern "C" from the dll-exports in the GameEngine_*.h files (but left them in the main.cpp). Intellisense in VS 2010 now knows these methods again as they are real C++ exports. WARNING: You definitely have to rebuild your GameEngine application. Maybe you additionally have to delete  the temporary files in the lib/include/bin/Build folder by hand, too.
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@752 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/src/GameAnimationComponent/KeyframeAnimComponent.cpp
@@ -444,7 +444,7 @@ void KeyframeAnimComponent::loadFromXml(const XMLNode* description)
 		{
 			// Find size of resource file
 			inf.seekg( 0, std::ios::end );
-			const int size = inf.tellg();
+			const int size = (int) inf.tellg();
 			// Copy resource file to memory
 			char *data = new char[size + 1];
 			inf.seekg( 0 );
